% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheeseSurvey4MCA.R, R/miniCheeseSurvey4MCA.R
\name{cheeseSurvey4MCA}
\alias{cheeseSurvey4MCA}
\title{An excel file for the data of
a survey questionnaire to be analyzed by MCA.}
\description{
\code{cheeseSurvey4MCA}:
An excel file
storing the data from an attitude survey:
305 (French) consumers
described their attitude, consumption, and knowledge about
a 
 cheese (Maroille) made in northen France.
 These data are analyzed with multiple correspondence analysis
 with the packages \code{ExPosition} and \code{PTCA4CATA}
(note that the package
\code{PTCA4CATA} needs to have been installed
from Github with the
command \code{devtools::install_github('HerveAbdi/PTCA4CATA')}).

\code{miniCheeseSurvey4MCA}:
An excel file
storing the data from an attitude survey:
305 (French) consumers
described their attitude, consumption, and knowledge about
 cheese (Maroille).
 These data are analyzed with multiple correspondence analysis
 with the packages \code{ExPosition} and \code{PTCA4CATA}
(note that the package
\code{PTCA4CATA} needs to have been installed
from github with the
command \code{devtools::install_github('HerveAbdi/PTCA4CATA')}).
}
\details{
The data are stored in an excel file containing 5  sheets.
The first sheet, called \code{Feuil1} contains the
 data for MCA: the rows are consumers and
columns are variables.
The other sheets (called \code{Feuil2} to \code{Feuil5})
provide additional information.
The questionnaire is provided as
\code{cheeseSurveyQuestionnaire}.

The data are stored in an excel file containing 5  sheets.
The first sheet, called \code{Feuil1} contains the
 data for MCA: the rows are consumers and
columns are variables.
The other sheets (called \code{Feuil2} to \code{Feuil5})
provide additional information.
The questionnaire is provided as
\code{cheeseSurveyQuestionnaire}.
The questions keept in the mini version concern
\emph{Knowledge} (8 questions), \emph{Beliefs} (24 questions)
along with descriptors of the participants: Age,
Sex, CSP, size of the family.
}
\section{FileName}{
 cheeseSurvey4MCA.xlsx

 miniCheeseSurvey4MCA.xlsx
}

\section{ReadingTheData}{

To fetch this dataset use \code{system.file()} (see example below).


To fetch this dataset use \code{system.file()} (see example below).
}

\examples{
\dontrun{
# Note we need to have DistatisR installed for the example to run
path2file <- system.file("extdata",
       "cheeseSurvey4MCA.xlsx", package = "R4SPISE2018")
cheeseDATA <- PTCA4CATA::read.xls.CATA(path = path2file,
                                         sheet = "Feuil1")
          }
\dontrun{
# Note we need to have DistatisR installed for the example to run
path2file <- system.file("extdata",
       "miniCheeseSurvey4MCA.xlsx", package = "R4SPISE2018")
cheeseDATA <- PTCA4CATA::read.xls.CATA(path = path2file,
                                         sheet = "Data")
          }
}
\references{
These data were used in a dissertation on
 Cheese consumption in France written by  Menouar Nacef
 under the direction of Professor Sylvie Chollet.

These data were used in a dissertation on
 Cheese consumption in France Menouar Nacef
 under the direction of Sylvie Chollet.
}
\author{
Nacef, M., & Chollet, S.

Nacef, M., & Chollet, S.
}
\keyword{DistatisR}
\keyword{R4SPISE2014}
\keyword{datasets}
