% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionsCheeseSurvey.R
\name{questionsCheeseSurvey}
\alias{questionsCheeseSurvey}
\title{A \code{docx} (Words) file
listing the questions used for
\code{cheeseSurvey4MCA.xlsx} and
\code{miniCheeseSurvey4MCA.xlsx}.}
\description{
\code{questionsCheeseSurvey}:
A \code{docx} (Words) file
listing the questions used for the survey
examples \code{cheeseSurvey4MCA.xlsx} and
\code{miniCheeseSurvey4MCA.xlsx}.
}
\details{
This is a standard \code{Words} file.
The questions are in the original French language,
an English version will come soon.
}
\section{FileName}{
 questionsCheeseSurvey.docx
}

\section{ReadingTheData}{

To fetch this dataset use \code{system.file()} (see example below).
}

\examples{
\dontrun{
# to get the path of the file:
path2file <- system.file("extdata",
       "questionsCheeseSurvey.docx", package = "R4SPISE2018")
          }
}
\references{
These data were used in a dissertation on
 Cheese consumption in France by Menouar Nacef
 under the direction of Professor Sylvie Chollet.
}
\author{
Nacef, M., & Chollet, S.
}
\keyword{DistatisR}
\keyword{datasets}
