% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4SPISE2018.R
\docType{package}
\name{R4SPISE2018}
\alias{R4SPISE2018}
\alias{R4SPISE2018-package}
\title{The R4SPISE2018 package provides utilities
(data sets, vignettes, scripts, etc.) for the SPISE 2018 Advanced
Workshop in Sensory Evaluation (Vietnam 07/23/2018 to 07/26/2018).}
\description{
The R4SPISE2018 package is
a compagnion of the SPISE2018
Advanced Workshop on Sensory Evaluation
(07/23/2018 to 07/26/2018, for details see:
  \url{http://sensorylab.hcmut.edu.vn/spise2018/}).
It provides
datasets, vignettes, scripts, and various utilities.
}
\author{
Herve Abdi \email{herve@utdallas.edu}
}
